# Conflicts:
#	gui/Pages/NumericalParamEquationView.py
#	src/gui/cs_gui_radiative_transfer.c
#	src/gui/cs_gui_specific_physics.c
#	src/lagr/cs_lagr_car.c
#	src/lagr/cs_lagr_options.c
#	src/lagr/cs_lagr_sde.c

+++ src/mesh/cs_join_mesh.c
@@ -3580,7 +3580,8 @@ cs_join_mesh_dump(FILE                  *f,
       cs_lnum_t  end = mesh->face_vtx_idx[i+1];
 
       fprintf(f, "\n face_id: %9ld gnum: %10llu n_vertices : %4ld\n",
-              (long)i, (unsigned long long)mesh->face_gnum[i], end-start);
+              (long)i, (unsigned long long)mesh->face_gnum[i],
+              (long)(end-start));
 
       for (j = start; j < end; j++) {
